Move padding helpers to shared module
Copilot-Session: 4c079ff0-a235-45de-b35d-154ca4812d34
